# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Player {
  id: number;
  name: string;
  createdAt: string;
}

export interface PlayerInput {
  /** @minLength 1 */
  name: string;
}

export interface PlayerUpdate {
  /** @minLength 1 */
  name?: string;
}

export interface PlayerSummary {
  playerId: number;
  playerName: string;
  games: number;
  wins: number;
  losses: number;
  points: number;
  rebounds: number;
  assists: number;
  steals: number;
  turnovers: number;
  blocks: number;
  ftMade: number;
  ftAttempted: number;
  twoMade: number;
  twoAttempted: number;
  threeMade: number;
  threeAttempted: number;
  ppg: number;
  rpg: number;
  apg: number;
  spg: number;
  topg: number;
  bpg: number;
  fgPct: number;
  threePct: number;
  ftPct: number;
}

export interface PlayerTeamGroup {
  teamId: number;
  teamName: string;
  games: number;
  wins: number;
  losses: number;
}

export interface Team {
  id: number;
  name: string;
  createdAt: string;
}

export interface TeamInput {
  /** @minLength 1 */
  name: string;
}

export interface TeamUpdate {
  /** @minLength 1 */
  name?: string;
}

export interface PlayerGameStatLine {
  playerId: number;
  playerName: string;
  ftMade: number;
  ftAttempted: number;
  twoMade: number;
  twoAttempted: number;
  threeMade: number;
  threeAttempted: number;
  points: number;
  assists: number;
  rebounds: number;
  steals: number;
  turnovers: number;
  blocks: number;
}

export interface PlayerGameStatInput {
  playerId: number;
  /** @minimum 0 */
  ftMade: number;
  /** @minimum 0 */
  ftAttempted: number;
  /** @minimum 0 */
  twoMade: number;
  /** @minimum 0 */
  twoAttempted: number;
  /** @minimum 0 */
  threeMade: number;
  /** @minimum 0 */
  threeAttempted: number;
  /** @minimum 0 */
  assists: number;
  /** @minimum 0 */
  rebounds: number;
  /** @minimum 0 */
  steals: number;
  /** @minimum 0 */
  turnovers: number;
  /** @minimum 0 */
  blocks: number;
}

export type GameResult = typeof GameResult[keyof typeof GameResult];


export const GameResult = {
  W: 'W',
  L: 'L',
} as const;

export type GameHighlightStatus = typeof GameHighlightStatus[keyof typeof GameHighlightStatus] | null;


export const GameHighlightStatus = {
  idle: 'idle',
  processing: 'processing',
  ready: 'ready',
  failed: 'failed',
} as const;

export interface GameEvent {
  playerId: number;
  statField: string;
  delta: number;
  /** @minimum 0 */
  videoTimestampMs: number;
}

export interface Game {
  id: number;
  teamId: number;
  teamName: string;
  opponent: string;
  date: string;
  result: GameResult;
  teamScore: number;
  opponentScore: number;
  videoObjectPath?: string | null;
  highlightObjectPath?: string | null;
  highlightStatus?: GameHighlightStatus;
  highlightError?: string | null;
  createdAt: string;
  stats: PlayerGameStatLine[];
  events: GameEvent[];
}

export type GameInputResult = typeof GameInputResult[keyof typeof GameInputResult];


export const GameInputResult = {
  W: 'W',
  L: 'L',
} as const;

export interface GameInput {
  teamId: number;
  /** @minLength 1 */
  opponent: string;
  date: string;
  result: GameInputResult;
  /** @minimum 0 */
  teamScore: number;
  /** @minimum 0 */
  opponentScore: number;
  videoObjectPath?: string | null;
  stats: PlayerGameStatInput[];
  events: GameEvent[];
}

export type GameUpdateResult = typeof GameUpdateResult[keyof typeof GameUpdateResult];


export const GameUpdateResult = {
  W: 'W',
  L: 'L',
} as const;

export interface GameUpdate {
  teamId: number;
  /** @minLength 1 */
  opponent: string;
  date: string;
  result: GameUpdateResult;
  /** @minimum 0 */
  teamScore: number;
  /** @minimum 0 */
  opponentScore: number;
  videoObjectPath?: string | null;
  stats: PlayerGameStatInput[];
  events: GameEvent[];
}

export type ImportGameRowResult = typeof ImportGameRowResult[keyof typeof ImportGameRowResult];


export const ImportGameRowResult = {
  W: 'W',
  L: 'L',
} as const;

export interface ImportGameRow {
  teamName: string;
  opponent: string;
  date: string;
  result: ImportGameRowResult;
  teamScore: number;
  opponentScore: number;
  playerName: string;
  ftMade: number;
  ftAttempted: number;
  twoMade: number;
  twoAttempted: number;
  threeMade: number;
  threeAttempted: number;
  assists: number;
  rebounds: number;
  steals: number;
  turnovers: number;
  blocks: number;
}

export interface ImportInput {
  rows: ImportGameRow[];
}

export interface ImportResult {
  playersCreated: number;
  teamsCreated: number;
  gamesCreated: number;
  statLinesCreated: number;
}

export interface UploadUrlRequest {
  /**
     * Original file name.
     * @minLength 1
     */
  name: string;
  /**
     * File size in bytes.
     * @minimum 1
     */
  size: number;
  /**
     * MIME type of the file (e.g. `video/webm`).
     * @minLength 1
     */
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database. */
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

/**
 * Effective plan derived from Stripe subscription status.
 */
export type BillingStatusPlan = typeof BillingStatusPlan[keyof typeof BillingStatusPlan];


export const BillingStatusPlan = {
  free: 'free',
  pro: 'pro',
} as const;

export interface BillingStatus {
  /** Effective plan derived from Stripe subscription status. */
  plan: BillingStatusPlan;
  /** Raw Stripe subscription status (e.g. trialing, active, past_due, canceled), or null if never subscribed. */
  status: string | null;
  currentPeriodEnd?: string | null;
  trialEnd?: string | null;
  cancelAtPeriodEnd: boolean;
}

/**
 * Which Pro price to check out with.
 */
export type CreateCheckoutSessionInputInterval = typeof CreateCheckoutSessionInputInterval[keyof typeof CreateCheckoutSessionInputInterval];


export const CreateCheckoutSessionInputInterval = {
  month: 'month',
  year: 'year',
} as const;

export interface CreateCheckoutSessionInput {
  /** Which Pro price to check out with. */
  interval: CreateCheckoutSessionInputInterval;
}

export interface CheckoutSessionResponse {
  /** Stripe-hosted Checkout URL to redirect the user to. */
  url: string;
}

export interface PortalSessionResponse {
  /** Stripe-hosted Billing Portal URL to redirect the user to. */
  url: string;
}

export type HighlightStatusStatus = typeof HighlightStatusStatus[keyof typeof HighlightStatusStatus];


export const HighlightStatusStatus = {
  idle: 'idle',
  processing: 'processing',
  ready: 'ready',
  failed: 'failed',
} as const;

export interface HighlightStatus {
  status: HighlightStatusStatus;
  highlightObjectPath?: string | null;
  error?: string | null;
  /** Number of qualifying moments (made shots, rebounds, assists, steals, blocks). */
  eligibleMoments: number;
}