# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games:
    post:
      operationId: createGame
      tags: [games]
      summary: Create a game with per-player stat lines
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameInput"
      responses:
        "201":
          description: Created game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{gameId}:
    parameters:
      - name: gameId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getGame
      tags: [games]
      summary: Get a single game with all player stat lines
      responses:
        "200":
          description: Game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
    patch:
      operationId: updateGame
      tags: [games]
      summary: Update a game and its per-player stat lines
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameUpdate"
      responses:
        "200":
          description: Updated game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
    delete:
      operationId: deleteGame
      tags: [games]
      summary: Delete a game
      responses:
        "204":
          description: Deleted
...
  /billing/checkout:
    post:
      operationId: createCheckoutSession
      tags: [billing]
      summary: Create a Stripe Checkout session to start/upgrade a Pro subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutSessionInput"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
        "400":
          description: Invalid price selection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /billing/portal:
    post:
      operationId: createBillingPortalSession
      tags: [billing]
      summary: Create a Stripe Billing Portal session for the caller to manage their subscription
      responses:
        "200":
          description: Billing portal session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalSessionResponse"
        "404":
          description: No Stripe customer exists for this account yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    PlayerGameStatLine:
      type: object
      required:
        - playerId
        - playerName
        - ftMade
        - ftAttempted
        - twoMade
        - twoAttempted
        - threeMade
        - threeAttempted
        - points
        - assists
        - rebounds
        - steals
        - turnovers
        - blocks
      properties:
        playerId: { type: integer }
        playerName: { type: string }
        ftMade: { type: integer }
        ftAttempted: { type: integer }
        twoMade: { type: integer }
        twoAttempted: { type: integer }
        threeMade: { type: integer }
        threeAttempted: { type: integer }
        points: { type: integer }
        assists: { type: integer }
        rebounds: { type: integer }
        steals: { type: integer }
        turnovers: { type: integer }
        blocks: { type: integer }
    PlayerGameStatInput:
      type: object
      required:
        - playerId
        - ftMade
        - ftAttempted
        - twoMade
        - twoAttempted
        - threeMade
        - threeAttempted
        - assists
        - rebounds
        - steals
        - turnovers
        - blocks
      properties:
        playerId: { type: integer }
        ftMade: { type: integer, minimum: 0 }
        ftAttempted: { type: integer, minimum: 0 }
        twoMade: { type: integer, minimum: 0 }
        twoAttempted: { type: integer, minimum: 0 }
        threeMade: { type: integer, minimum: 0 }
        threeAttempted: { type: integer, minimum: 0 }
        assists: { type: integer, minimum: 0 }
        rebounds: { type: integer, minimum: 0 }
        steals: { type: integer, minimum: 0 }
        turnovers: { type: integer, minimum: 0 }
        blocks: { type: integer, minimum: 0 }
    Game:
      type: object
      required:
        - id
        - teamId
        - teamName
        - opponent
        - date
        - result
        - teamScore
        - opponentScore
        - createdAt
        - stats
        - events
      properties:
        id: { type: integer }
        teamId: { type: integer }
        teamName: { type: string }
        opponent: { type: string }
        date: { type: string, format: date }
        result: { type: string, enum: [W, L] }
        teamScore: { type: integer }
        opponentScore: { type: integer }
        videoObjectPath: { type: string, nullable: true }
        highlightObjectPath: { type: string, nullable: true }
        highlightStatus:
          type: string
          nullable: true
          enum: [idle, processing, ready, failed]
        highlightError: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        stats:
          type: array
          items:
            $ref: "#/components/schemas/PlayerGameStatLine"
        events:
          type: array
          items:
            $ref: "#/components/schemas/GameEvent"
    GameInput:
      type: object
      required: [teamId, opponent, date, result, teamScore, opponentScore, stats, events]
      properties:
        teamId: { type: integer }
        opponent: { type: string, minLength: 1 }
        date: { type: string, format: date }
        result: { type: string, enum: [W, L] }
        teamScore: { type: integer, minimum: 0 }
        opponentScore: { type: integer, minimum: 0 }
        videoObjectPath: { type: string, nullable: true }
        stats:
          type: array
          items:
            $ref: "#/components/schemas/PlayerGameStatInput"
        events:
          type: array
          items:
            $ref: "#/components/schemas/GameEvent"
    GameUpdate:
      type: object
      required: [teamId, opponent, date, result, teamScore, opponentScore, stats, events]
      properties:
        teamId: { type: integer }
        opponent: { type: string, minLength: 1 }
        date: { type: string, format: date }
        result: { type: string, enum: [W, L] }
        teamScore: { type: integer, minimum: 0 }
        opponentScore: { type: integer, minimum: 0 }
        videoObjectPath: { type: string, nullable: true }
        stats:
          type: array
          items:
            $ref: "#/components/schemas/PlayerGameStatInput"
        events:
          type: array
          items:
            $ref: "#/components/schemas/GameEvent"
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CreateCheckoutSessionInput:
      type: object
      required: [interval]
      properties:
        interval:
          type: string
          enum: [month, year]
          description: Which Pro price to check out with.
    CheckoutSessionResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: Stripe-hosted Checkout URL to redirect the user to.
    PortalSessionResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: Stripe-hosted Billing Portal URL to redirect the user to.
...
    GameEvent:
      type: object
      required: [playerId, statField, delta, videoTimestampMs]
      properties:
        playerId: { type: integer }
        statField: { type: string }
        delta: { type: integer }
        videoTimestampMs: { type: integer, minimum: 0 }