# lib/api-zod/src/generated/types/checkoutSessionResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutSessionResponse {
  /** Stripe-hosted Checkout URL to redirect the user to. */
  url: string;
}