# lib/api-zod/src/generated/types/createCheckoutSessionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCheckoutSessionInputInterval } from './createCheckoutSessionInputInterval';

export interface CreateCheckoutSessionInput {
  /** Which Pro price to check out with. */
  interval: CreateCheckoutSessionInputInterval;
}