# lib/api-zod/src/generated/types/portalSessionResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PortalSessionResponse {
  /** Stripe-hosted Billing Portal URL to redirect the user to. */
  url: string;
}